// src/abis/types/PoolManager.ts
export type PoolKeyStruct = {
  currency0: string;
  currency1: string;
  fee: BigNumberish;
  tickSpacing: BigNumberish;
  hooks: string;
};